third_party/WebKit/LayoutTests/external/wpt/css/css-transforms/animation/list-interpolation-expected.txt

This is a testharness.js-based test.
PASS none -> none: "none" and "none" are valid transform values
FAIL none -> none: Animation between "none" and "none" at progress 0.25 assert_equals: expected "none" but got "matrix(1, 0, 0, 1, 0, 0)"
PASS none -> something: "none" and "translate(200px) rotate(720deg)" are valid transform values
PASS none -> something: Animation between "none" and "translate(200px) rotate(720deg)" at progress 0.25
PASS something -> none: "translate(200px) rotate(720deg)" and "none" are valid transform values
PASS something -> none: Animation between "translate(200px) rotate(720deg)" and "none" at progress 0.25
PASS Mismatched lengths (from is shorter), common part matches: "translate(100px)" and "translate(200px) rotate(720deg)" are valid transform values
FAIL Mismatched lengths (from is shorter), common part matches: Animation between "translate(100px)" and "translate(200px) rotate(720deg)" at progress 0.25 assert_equals: expected "matrix(-1, 1.22465e-16, -1.22465e-16, -1, 125, 0)" but got "matrix(1, 0, 0, 1, 125, 0)"
PASS Mismatched lengths (to is shorter), common part matches: "translate(100px) rotate(720deg)" and "translate(200px)" are valid transform values
FAIL Mismatched lengths (to is shorter), common part matches: Animation between "translate(100px) rotate(720deg)" and "translate(200px)" at progress 0.25 assert_equals: expected "matrix(-1, 3.67394e-16, -3.67394e-16, -1, 125, 0)" but got "matrix(1, -4.89859e-16, 4.89859e-16, 1, 125, 0)"
PASS Perfect match: "scale(2) rotate(360deg) translate(100px) matrix(1, 0, 0, 1, 100, 0) skew(0deg)" and "scale(3) rotate(1080deg) translate(200px) matrix(1, 0, 0, 1, 0, 200) skew(720deg)" are valid transform values
PASS Perfect match: Animation between "scale(2) rotate(360deg) translate(100px) matrix(1, 0, 0, 1, 100, 0) skew(0deg)" and "scale(3) rotate(1080deg) translate(200px) matrix(1, 0, 0, 1, 0, 200) skew(720deg)" at progress 0.25
PASS Matches on primitives: "translateX(100px) scaleX(3) translate(500px) scale(2)" and "translateY(200px) scale(5) translateX(100px) scaleY(3)" are valid transform values
PASS Matches on primitives: Animation between "translateX(100px) scaleX(3) translate(500px) scale(2)" and "translateY(200px) scale(5) translateX(100px) scaleY(3)" at progress 0.25
PASS Match on rotation vector: "rotateX(90deg) translateX(100px)" and "rotate3d(50, 0, 0, 180deg) translateY(200px)" are valid transform values
FAIL Match on rotation vector: Animation between "rotateX(90deg) translateX(100px)" and "rotate3d(50, 0, 0, 180deg) translateY(200px)" at progress 0.25 assert_equals: expected "matrix3d(1, 0, 0, 0, 0, -0.382683, 0.92388, 0, 0, -0.92388, -0.382683, 0, 75, -19.1342, 46.194, 1)" but got "matrix3d(1, 0, 0, 0, 0, -1, 1.22465e-16, 0, 0, -1.22465e-16, -1, 0, 75, -50, 6.12323e-15, 1)"
PASS Match on rotation due to 0deg angle: "rotateX(90deg) translateX(100px)" and "rotateY(0deg) translateY(200px)" are valid transform values
FAIL Match on rotation due to 0deg angle: Animation between "rotateX(90deg) translateX(100px)" and "rotateY(0deg) translateY(200px)" at progress 0.25 assert_equals: expected "matrix3d(1, 0, 0, 0, 0, 0.382683, 0.92388, 0, 0, -0.92388, 0.382683, 0, 75, 19.1342, 46.194, 1)" but got "matrix(1, 0, 0, 1, 75, 50)"
PASS Common prefix: "rotate(0deg) translate(100px)" and "rotate(720deg) scale(2) translate(200px)" are valid transform values
FAIL Common prefix: Animation between "rotate(0deg) translate(100px)" and "rotate(720deg) scale(2) translate(200px)" at progress 0.25 assert_equals: expected "matrix(-1.25, 1.53081e-16, -1.53081e-16, -1.25, -175, 2.14313e-14)" but got "matrix(1.25, 0, 0, 1.25, 175, -4.89859e-14)"
PASS Complete mismatch (except length): "scale(2) rotate(0deg) translate(100px)" and "rotate(720deg) scale(2) translate(200px)" are valid transform values
FAIL Complete mismatch (except length): Animation between "scale(2) rotate(0deg) translate(100px)" and "rotate(720deg) scale(2) translate(200px)" at progress 0.25 assert_equals: expected "matrix(2, 0, 0, 2, 250, 0)" but got "matrix(2, 0, 0, 2, 250, -4.89859e-14)"
PASS Complete mismatch including length: "scale(2) rotate(0deg)" and "rotate(720deg) scale(2) translate(200px)" are valid transform values
FAIL Complete mismatch including length: Animation between "scale(2) rotate(0deg)" and "rotate(720deg) scale(2) translate(200px)" at progress 0.25 assert_equals: expected "matrix(2, 0, 0, 2, 100, 0)" but got "matrix(2, 0, 0, 2, 100, -4.89859e-14)"
Harness: the test ran to completion.
